+++ src/main/java/org/bukkit/World.java
@@ -1,6 +1,7 @@
 
 package org.bukkit;
 
+import org.bukkit.block.Block;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.entity.ItemDrop;
 import org.bukkit.entity.PoweredMinecart;
