+++ src/main/java/org/bukkit/block/Biome.java
@@ -1,5 +1,5 @@
 
-package org.bukkit;
+package org.bukkit.block;
 
 /**
  * Holds all accepted Biomes in the default server
