+++ src/main/java/org/bukkit/block/Block.java
@@ -1,5 +1,8 @@
-package org.bukkit;
+package org.bukkit.block;
 
+import org.bukkit.Chunk;
+import org.bukkit.Material;
+import org.bukkit.World;
 import org.bukkit.block.BlockState;
 
 /**
