+++ src/main/java/org/bukkit/block/BlockFace.java
@@ -1,4 +1,4 @@
-package org.bukkit;
+package org.bukkit.block;
 
 /**
  * Represents the face of a block
