+++ src/main/java/org/bukkit/event/block/BlockCanBuildEvent.java
@@ -3,7 +3,7 @@
  */
 package org.bukkit.event.block;
 
-import org.bukkit.Block;
+import org.bukkit.block.Block;
 import org.bukkit.Material;
 
 /**
