+++ src/main/java/org/bukkit/event/block/BlockIgniteEvent.java
@@ -1,6 +1,6 @@
 package org.bukkit.event.block;
 
-import org.bukkit.Block;
+import org.bukkit.block.Block;
 import org.bukkit.entity.Player;
 import org.bukkit.event.Cancellable;
 import org.bukkit.event.Event;
