+++ src/main/java/org/bukkit/event/block/BlockPhysicsEvent.java
@@ -1,6 +1,6 @@
 package org.bukkit.event.block;
 
-import org.bukkit.Block;
+import org.bukkit.block.Block;
 import org.bukkit.Material;
 import org.bukkit.event.Event;
