+++ src/main/java/org/bukkit/event/block/LeavesDecayEvent.java
@@ -1,7 +1,7 @@
 
 package org.bukkit.event.block;
 
-import org.bukkit.Block;
+import org.bukkit.block.Block;
 import org.bukkit.event.Cancellable;
 
 /**
