+++ src/main/java/org/bukkit/material/Ladder.java
@@ -1,7 +1,7 @@
 
 package org.bukkit.material;
 
-import org.bukkit.BlockFace;
+import org.bukkit.block.BlockFace;
 import org.bukkit.Material;
 
 /**
